+++ tests/integration-tests/tests/common/utils.py
@@ -47,6 +47,7 @@ NODE_TYPES = (HEAD_NODE, COMPUTE_NODE, LOGIN_NODE)
 GPU_JOB_SCRIPT = pathlib.Path(__file__).parent / "data/gpu_job.sh"
 
 RHEL_OWNERS = ["309956199498", "841258680906", "219670896067"]
+ALMALINUX_OWNERS = ["764336703387"]
 
 OS_TO_OFFICIAL_AMI_NAME_OWNER_MAP = {
     "alinux2023": {"name": "al2023-ami-2023.*.*.*-kernel-6.1-*", "owners": ["amazon"]},
@@ -73,6 +74,10 @@ OS_TO_OFFICIAL_AMI_NAME_OWNER_MAP = {
     "rocky8.9": {"name": "Rocky-8-EC2-Base-8.9*", "owners": ["792107900819"]},  # TODO add china and govcloud accounts
     "rhel9": {"name": "RHEL-9.*_HVM*", "owners": RHEL_OWNERS},
     "rocky9": {"name": "Rocky-9-EC2-Base-9.*", "owners": ["792107900819"]},  # TODO add china and govcloud accounts
+    "almalinux8": {
+        "name": {"x86_64": "AlmaLinux OS 8*x86_64", "arm64": "AlmaLinux OS 8*aarch64"},
+        "owners": ALMALINUX_OWNERS,
+    },
 }
 
 # Remarkable AMIs are latest deep learning base AMI and FPGA developer AMI without pcluster infrastructure
@@ -120,6 +125,7 @@ OS_TO_PCLUSTER_AMI_NAME_OWNER_MAP = {
     "rocky8": {"name": "rocky8-hvm-*-*", "owners": PCLUSTER_AMI_OWNERS},
     "rhel9": {"name": "rhel9-hvm-*-*", "owners": PCLUSTER_AMI_OWNERS},
     "rocky9": {"name": "rocky9-hvm-*-*", "owners": PCLUSTER_AMI_OWNERS},
+    "almalinux8": {"name": "almalinux8-hvm-*-*", "owners": PCLUSTER_AMI_OWNERS},
 }
 
 FIRST_STAGE_AMI_OWNERS = ["self", "447714826191"]